Fixed tests and build errors
