+++ CoBySi.Pomodoro.Web/Components/Pages/Home.razor
@@ -1,3 +1,3 @@
 ﻿@page "/"
 <PageTitle>Home</PageTitle>
-<Pomodoro></Pomodoro>
+<PomodoroComponent></PomodoroComponent>
