+++ Tests/CoBySi.Pomodoro.Tests/lcov.info
@@ -22,27 +22,29 @@ SF:c:\Users\simon\Documents\projects\pomodoroTimer\CoBySi.Pomodoro\PomodorHandle
 FN:9,System.Nullable`1<System.Double> CoBySi.Pomodoro.PomodorHandler::get__totalNumberOfSeconds()
 FNDA:0,System.Nullable`1<System.Double> CoBySi.Pomodoro.PomodorHandler::get__totalNumberOfSeconds()
 DA:10,0
-FN:23,System.Void CoBySi.Pomodoro.PomodorHandler::Start(CoBySi.Pomodoro.Timer.PomodoroState)
-FNDA:0,System.Void CoBySi.Pomodoro.PomodorHandler::Start(CoBySi.Pomodoro.Timer.PomodoroState)
+FN:21,System.Void CoBySi.Pomodoro.PomodorHandler::Start(CoBySi.Pomodoro.Timer.PomodoroState,System.Nullable`1<System.Double>)
+FNDA:0,System.Void CoBySi.Pomodoro.PomodorHandler::Start(CoBySi.Pomodoro.Timer.PomodoroState,System.Nullable`1<System.Double>)
+DA:22,0
+DA:23,0
 DA:24,0
 DA:25,0
 DA:26,0
 DA:27,0
 DA:28,0
 DA:29,0
-DA:30,0
-DA:31,0
-BRDA:29,95,0,0
-BRDA:29,95,1,0
-BRDA:29,129,0,0
-BRDA:29,129,1,0
-FN:32,System.Void CoBySi.Pomodoro.PomodorHandler::TimerCallback(System.Object)
+BRDA:27,84,0,0
+BRDA:27,84,1,0
+BRDA:27,118,0,0
+BRDA:27,118,1,0
+FN:30,System.Void CoBySi.Pomodoro.PomodorHandler::TimerCallback(System.Object)
 FNDA:0,System.Void CoBySi.Pomodoro.PomodorHandler::TimerCallback(System.Object)
+DA:31,0
+DA:32,0
 DA:33,0
-DA:34,0
-DA:35,0
-FN:37,System.Void CoBySi.Pomodoro.PomodorHandler::Tick()
+FN:35,System.Void CoBySi.Pomodoro.PomodorHandler::Tick()
 FNDA:0,System.Void CoBySi.Pomodoro.PomodorHandler::Tick()
+DA:36,0
+DA:37,0
 DA:38,0
 DA:39,0
 DA:40,0
@@ -51,61 +53,44 @@ DA:42,0
 DA:43,0
 DA:44,0
 DA:45,0
-DA:46,0
-DA:48,0
-BRDA:40,22,0,0
-BRDA:40,22,1,0
-BRDA:40,56,0,0
-BRDA:40,56,1,0
-BRDA:41,141,0,0
-BRDA:44,158,0,0
-BRDA:44,158,1,0
-BRDA:41,141,1,0
-FN:50,System.Void CoBySi.Pomodoro.PomodorHandler::StopTimer()
+BRDA:38,22,0,0
+BRDA:38,22,1,0
+BRDA:38,56,0,0
+BRDA:38,56,1,0
+BRDA:39,141,0,0
+BRDA:42,158,0,0
+BRDA:42,158,1,0
+BRDA:39,141,1,0
+FN:47,System.Void CoBySi.Pomodoro.PomodorHandler::StopTimer()
 FNDA:0,System.Void CoBySi.Pomodoro.PomodorHandler::StopTimer()
-DA:51,0
-DA:52,0
+DA:48,0
+DA:49,0
+DA:50,0
+BRDA:49,8,0,0
+BRDA:49,8,1,0
+FN:52,System.Void CoBySi.Pomodoro.PomodorHandler::Stop(System.Nullable`1<System.Double>)
+FNDA:0,System.Void CoBySi.Pomodoro.PomodorHandler::Stop(System.Nullable`1<System.Double>)
 DA:53,0
-BRDA:52,8,0,0
-BRDA:52,8,1,0
-FN:55,System.Nullable`1<System.Double> CoBySi.Pomodoro.PomodorHandler::GetTotalNumberOfSeconds(CoBySi.Pomodoro.Timer.PomodoroState,CoBySi.Pomodoro.PomodoroSettings)
-FNDA:0,System.Nullable`1<System.Double> CoBySi.Pomodoro.PomodorHandler::GetTotalNumberOfSeconds(CoBySi.Pomodoro.Timer.PomodoroState,CoBySi.Pomodoro.PomodoroSettings)
+DA:54,0
+DA:55,0
 DA:56,0
 DA:57,0
 DA:58,0
-DA:59,0
-DA:60,0
-DA:61,0
-DA:62,0
-DA:63,0
-BRDA:57,7,0,0
-BRDA:57,7,1,0
-BRDA:57,13,1,0
-BRDA:57,13,0,0
-FN:65,System.Void CoBySi.Pomodoro.PomodorHandler::Stop()
-FNDA:0,System.Void CoBySi.Pomodoro.PomodorHandler::Stop()
-DA:66,0
-DA:67,0
-DA:68,0
-DA:69,0
-DA:70,0
-DA:71,0
-BRDA:70,46,0,0
-BRDA:70,46,1,0
-BRDA:70,80,0,0
-BRDA:70,80,1,0
-FN:16,System.Void CoBySi.Pomodoro.PomodorHandler::.ctor(CoBySi.Pomodoro.PomodoroSettings,System.TimeProvider)
-FNDA:0,System.Void CoBySi.Pomodoro.PomodorHandler::.ctor(CoBySi.Pomodoro.PomodoroSettings,System.TimeProvider)
+BRDA:57,30,0,0
+BRDA:57,30,1,0
+BRDA:57,59,0,0
+BRDA:57,59,1,0
+FN:15,System.Void CoBySi.Pomodoro.PomodorHandler::.ctor(System.TimeProvider)
+FNDA:0,System.Void CoBySi.Pomodoro.PomodorHandler::.ctor(System.TimeProvider)
+DA:16,0
 DA:17,0
 DA:18,0
 DA:19,0
-DA:20,0
-DA:21,0
-LF:44
+LF:35
 LH:0
-BRF:22
+BRF:18
 BRH:0
-FNF:8
+FNF:7
 FNH:0
 end_of_record
 SF:c:\Users\simon\Documents\projects\pomodoroTimer\CoBySi.Pomodoro\Timer\TimeChangedEventArgs.cs
